# cmd/main/docs/swagger.yaml
basePath: /
definitions:
  gin.H:
    additionalProperties: {}
    type: object
  models.Todo:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      title:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Todo API Documentation. Written in Go. Gin used
  title: Todo API
  version: "1.0"
paths:
  /todo:
    post:
      description: Add a todo to the DB
      parameters:
      - description: Todo object that is to be created
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/models.Todo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Todo'
      summary: Create todo
      tags:
      - todo
  /todo/:id:
    delete:
      description: Delete a todo from the DB
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/gin.H'
      summary: Delete todo
      tags:
      - todo
    put:
      description: Update a todo in the DB
      parameters:
      - description: Todo object that is to be updated
        in: body
        name: todo
        required: true
        schema:
          $ref: '#/definitions/models.Todo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Todo'
      summary: Update todo
      tags:
      - todo
  /todos:
    get:
      description: Fetch todos from DB
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Todo'
      summary: Get todos
      tags:
      - todo
swagger: "2.0"